Crossed snek and stacked droop
Snek crossing is now allowed as long a the heights are far enough apart.
Collision is in place for wall collision, and self collision works provided that air control is not allowed, a colliding with a snek tile after the head but before the last grounded snek piece won't droop. This doesn't matter as air control will always be off except for debug.
To do: add script to set the vairable SnekInAir, used to toggle air control.